computer-science/nlp/_src/llama_index_local/pyproject.toml
[tool.poetry]
name = "llama_index-local"
version = "0.1.0"
description = ""
authors = ["xhiroga <xhiroga@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "llama_index_local"}]

[tool.poetry.dependencies]
python = "3.9.12"
python-dotenv = "^1.0.0"
httpx = "^0.24.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
flake8-bugbear = "^23.3.23"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
